Additional Tcl search path
I don't know if it's due to the GitHub macos-latest build runner image being
macos 14 or due to the update to Tcl 8.6.15, but the Tcl installation now
appears under /System/Library/Frameworks/Tcl.framework/Versions/, which we
weren't searching before, leading to build failures. Add this path to the
list of system Tcl paths to check. Hopefully this fixes the build 🤞.