Find correct brew-installed Tcl
Revert the last commit, which didn't help find the correct Tcl installation.

Since the default brew-installed version of Tcl is now 9.0, the runner image
installs Tcl 8 using the prefix `tcl-tk@8`. Try both prefixes and use the one
that works.

Opened #524, as well, to consider whether we need to support Tcl 9 in the
future.